requirements.txt
# Model Training
datasets
transformers[torch]
huggingface_hub
pandas
tqdm
nltk
gradio
tensorboard
numpy<2.0
# Linebot
whisper
jieba
line-bot-sdk

# torch (See https://pytorch.org/ and choose version that match your GPU.)
# CUDA 11.8 -> pip install torch==2.3.1 torchvision==0.18.1 torchaudio==2.3.1 --index-url https://download.pytorch.org/whl/cu118
# CUDA 12.1 -> pip install torch==2.3.1 torchvision==0.18.1 torchaudio==2.3.1 --index-url https://download.pytorch.org/whl/cu121

# Others
pillow
matplotlib
wordcloud
openai
Flask
scikit-learn
tiktoken
pyodbc
flask_sqlalchemy
python-dotenv